カジュアルマッチ開始を別ファイルに切り出した
